Refactors and sonar issues addressed.
